Attempting to fix battery kill bug (1)
I've noticed that the battery in my journey has been dead a few times more than I care to happen. After watching my vehicle, i realized that the center console on my dashboard turns on every few seconds as though something isn't letting the ECU sleep correctly. I'm not sure if the scan port is causing it or if the RPM check is causing it, but this is an attempt (1) to resolve it.